fix(test/tap): batch 3 - remaining narrowing fixes in admin tests and anomaly_detector_unit

+++ test/tap/tests/admin_show_create_table-t.cpp
@@ -70,9 +70,9 @@ int main() {
 	}
 	mysql_close(proxysql_admin);
 	if (cl.use_noise) {
-		plan(tables.size() + 1 + 2);
+		plan(static_cast<int>(tables.size()) + 1 + 2);
 	} else {
-		plan(tables.size() + 1);
+		plan(static_cast<int>(tables.size()) + 1);
 	}
 	ok(tables.size() > 40 , "Number of tables to check: %ld" , tables.size());
